Charge a configuration what it costs, and bound it by one number
A device had four bounds: a fixed sixty-four byte slot per object however
much it held, a count of objects, a count of each kind, and a declared
byte capacity nothing enforced. The K1 reserved 3,220 bytes across its two
snapshots to hold at most about 880, and refused projects for counts it had
the room for.

The store is now a packed byte arena. Objects live end to end as
(kind, id, length, payload) entries in canonical key order, which is
exactly what a configuration image carries after its header, so retaining
the active snapshot is a copy rather than a sorted rebuild, and the shared
copy the interface reads is the same bytes again. Decoders take a borrowed
object, so nothing is copied into a worst-case buffer to be read.

What a device declares is configuration_bytes and everything else follows
from it: max_objects is the count those bytes imply, and a host refuses a
project for the bytes it needs rather than for how many objects they are.
MAX_CHANNELS, MAX_BANKS, MAX_GENERATED_BANKS and kind_limits are gone. The
K1 declares 1,264 bytes, which buys about twenty-six explicit channels where
eight were allowed, or as many plans as fit — and static RAM fell from 9,284
bytes to 8,188 with 8,196 free.

A generated bank is now a shared core plus its own encoding's tail, and the
encoding is a property the plan declares rather than an inference from a
zero offset. A simplex band stores 56 bytes and a repeater sub-band 60,
where both stored 59 before. Storage format version 4.

+++ crates/radio-firmware-k1/src/shell.rs
@@ -15,8 +15,8 @@
 
 use radio_domain::{BankId, SquelchLevel, MAX_SQUELCH_LEVEL};
 
-use crate::configuration::MAX_BANKS;
 use crate::keypad::Key;
+use radio_channel_plan::MAX_BANKS;
 
 /// Digits a channel-number entry accepts.
 pub const ENTRY_DIGITS: usize = 2;
@@ -183,7 +183,7 @@ pub struct Shell {
     cursor: u16,
     entry: Option<Entry>,
     bank_filter: Option<BankId>,
-    banks: [Option<BankId>; MAX_BANKS],
+    banks: [Option<BankId>; MAX_BANKS as usize],
     bank_count: usize,
     source_cursor: usize,
     vfo_hz: u32,
@@ -212,7 +212,7 @@ impl Shell {
             cursor: 0,
             entry: None,
             bank_filter: None,
-            banks: [None; MAX_BANKS],
+            banks: [None; MAX_BANKS as usize],
             bank_count: 0,
             source_cursor: 0,
             vfo_hz: VFO_DEFAULT_HZ,
@@ -362,9 +362,9 @@ impl Shell {
     ///
     /// A filter which the new configuration does not populate is cleared, so
     /// the operator can never be left looking at an empty view.
-    pub fn set_banks(&mut self, banks: [Option<BankId>; MAX_BANKS], count: usize) -> bool {
+    pub fn set_banks(&mut self, banks: [Option<BankId>; MAX_BANKS as usize], count: usize) -> bool {
         self.banks = banks;
-        self.bank_count = count.min(MAX_BANKS);
+        self.bank_count = count.min(MAX_BANKS as usize);
         self.entry = None;
         self.cursor = 0;
         if self
@@ -764,8 +764,8 @@ mod tests {
         Context, Intent, Mode, Screen, Shell, Source, ENTRY_TIMEOUT_MILLISECONDS, VFO_DEFAULT_HZ,
         VFO_MAXIMUM_HZ, VFO_STEPS_HZ,
     };
-    use crate::configuration::MAX_BANKS;
     use crate::keypad::Key;
+    use radio_channel_plan::MAX_BANKS;
     use radio_domain::{BankId, SquelchLevel};
 
     fn context(visible: u16, active: u16) -> Context {
@@ -775,8 +775,8 @@ mod tests {
         }
     }
 
-    fn bank_table(ids: &[u16]) -> ([Option<BankId>; MAX_BANKS], usize) {
-        let mut banks = [None; MAX_BANKS];
+    fn bank_table(ids: &[u16]) -> ([Option<BankId>; MAX_BANKS as usize], usize) {
+        let mut banks = [None; MAX_BANKS as usize];
         for (slot, id) in banks.iter_mut().zip(ids) {
             *slot = Some(BankId::new(*id));
         }
